refactor: comprehensive codebase optimization — solver, reactivity, bundle, cleanup
40-finding analysis across WASM solver, TypeScript app, CSS, and build config.
Net result: ~1,000 lines removed, dramatically faster solver, smaller bundles.

Phase 1 — Quick wins:
- Cargo.toml: opt-level 3, wasm-opt with bulk-memory/nontrapping-float flags
- Remove dead components (ExportPanel, MarginalHistogram) and unused signals
- Remove orphaned CSS properties, empty rules, duplicate animations

Phase 2 — High-payoff performance:
- FISTA solver: FFT-based convolutions replacing O(n×k) time-domain loops
- Eliminate redundant 2nd forward convolution via primal residual convergence
- Float64→Float32 conversion moved to dispatch boundary (halves transfer size)
- Metrics: tab-visibility gate, status-based selective recompute, WeakMap SNR cache
- Supabase SDK: dynamic import() for code-splitting (~45KB saved when unused)
- Lazy tab rendering with mounted-set tracking

Phase 3 — Structural improvements:
- multi-cell-store: Map cloning → createStore with path-based setters
- SubmitPanel: 481→237 lines, extracted SubmitForm, GroundTruthControls, submitAction
- Import components moved to src/components/import/
- Import extensions standardized across 47 files
- WASM serialization: manual byte packing → Cursor-based reads

Phase 4 — Polish:
- ZoomWindow: 3-pass z-score stats → single pass
- Worker yield: setTimeout(0) → MessageChannel (<1ms vs ~4ms)
- Idiomatic .fill(0.0), dead code gated behind #[cfg(test)]
- CSS: fix mismatched color fallbacks, hardcoded colors → custom properties
- Google Fonts display=optional, @/ path aliases in tsconfig + vite

+++ src/lib/solver-types.ts
@@ -52,7 +52,7 @@ export type WarmStartStrategy = 'warm' | 'warm-no-momentum' | 'cold';
 
 /** Full solve request message sent to the worker. */
 export interface SolveRequest {
-  trace: Float64Array;
+  trace: Float32Array;
   params: SolverParams;
   warmStartState: Uint8Array | null;
   warmStartStrategy: WarmStartStrategy;
@@ -66,7 +66,7 @@ export type PoolWorkerInbound =
   | {
       type: 'solve';
       jobId: number;
-      trace: Float64Array;
+      trace: Float32Array;
       params: SolverParams;
       warmState: Uint8Array | null;
       warmStrategy: WarmStartStrategy;
